+++ TokanPages.ClientApp/src/app.tsx
@@ -4,8 +4,8 @@ import Fab from "@material-ui/core/Fab";
 import KeyboardArrowUpIcon from "@material-ui/icons/KeyboardArrowUp";
 import { GetContentManifestDto } from "./Api/Models";
 import { ScrollToTop } from "./Shared/Components/Scroll";
-import { UpdateUserData } from "./Shared/Services/updateUserDataService";
-import { UpdateUserLanguage } from "./Shared/Services/updateUserLanguageService";
+import { UpdateUserData } from "./Shared/Services/initializeService";
+import { UpdateUserLanguage } from "./Shared/Services/languageService";
 import { AppStyle } from "./app.style";
 import AOS from "aos";
 import "aos/dist/aos.css";
